chore: storybook fixes removed duplicate comp and some minor fixes
